api/requirements.txt
farm-haystack==1.21.0
flask==3.0.0
gunicorn==21.2.0
opensearch-py==2.3.2
python-dotenv==1.0.0
Flask-MySQLdb==2.0.0
sentence-transformers==2.2.2
wheel==0.38.4
openai>=1.2.3
